Reverse open issue logic (#37)
* reversing the logic for closing/opening issues

the current logic looks for currently opened (equivalent)
issues and closes them before opening a new issue. In
practice because review is slow (and the updater will find
changes and want to open a new issue) this is going to lead
to a HUGE number of opened / closed issues (depending on how
often it is run) so a better approach, implemented here, is
simply to check for open issues, and if one is found, do not
open a new one. We assume it would be redundant, and the
developer is aware that the previous one is open (and not
resolved yet).

Signed-off-by: vsoch <vsoch@users.noreply.github.com>